% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\name{bioclim_current}
\alias{bioclim_current}
\title{Current bioclimatic raster}
\format{
GeoTIFF file, readable with \code{terra::rast()}.
}
\description{
A GeoTIFF with current bioclimatic variables for Costa Rica.
}
\details{
This file is stored in \verb{inst/extdata/} and can be accessed with:
\code{terra::rast(system.file("extdata", "bioclim_current.tif", package = "h3sdm"))}
}
\examples{
library(terra)
bio <- terra::rast(system.file("extdata", "bioclim_current.tif", package = "h3sdm"))
}
