% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\name{bioclim_future}
\alias{bioclim_future}
\title{Future bioclimatic raster}
\format{
GeoTIFF file, readable with \code{terra::rast()}.
}
\description{
A GeoTIFF with projected bioclimatic variables for Costa Rica.
}
\details{
This dataset corresponds to the climate projection:
\itemize{
\item Model: INM-CM4-8
\item Scenario: SSP1-2.6
\item Period: 2021–2040
}

The file is stored in \verb{inst/extdata/} and can be accessed with:
\code{terra::rast(system.file("extdata", "bioclim_future.tif", package = "h3sdm"))}
}
\examples{
library(terra)
bio <- terra::rast(system.file("extdata", "bioclim_future.tif", package = "h3sdm"))
}
