% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3sdm_predict.R
\name{h3sdm_predict}
\alias{h3sdm_predict}
\title{Predict species distribution with H3 hexagons}
\usage{
h3sdm_predict(fit_object, new_data)
}
\arguments{
\item{new_data}{An \code{sf} object containing the spatial grid and
the same predictor variables used for model training.}

\item{sdm_workflow}{A fitted \code{tidymodels} workflow object,
typically the result of \code{workflows::fit()}.}
}
\value{
An \code{sf} object with a new column named \code{prediction} containing
the predicted probability of presence for each hexagon.
}
\description{
Uses a fitted tidymodels workflow to predict species
presence probability on a new spatial grid. This function
automatically handles the creation of 'x' and 'y' coordinates
if they are not present in the input data.
}
