% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3sdm_stack_blend.R
\name{h3sdm_stack_blend}
\alias{h3sdm_stack_blend}
\title{Crea un ensemble de modelos (Stack) y ajusta sus pesos.}
\usage{
h3sdm_stack_blend(..., non_negative = TRUE, metric = NULL)
}
\arguments{
\item{...}{Objetos de lista que son el resultado de \code{h3sdm_fit_model()}.
Cada objeto de lista debe contener el elemento \code{cv_model} (resultado de
\code{fit_resamples}). Se recomienda usar argumentos con nombre (ej., \code{rf = rf_fit}).}

\item{non_negative}{Lógico. Si es \code{TRUE} (por defecto), fuerza a que los
pesos de los modelos candidatos sean no negativos (contribución positiva).}

\item{metric}{La métrica utilizada para optimizar la combinación de pesos
durante el blending. Por defecto, usa la primera métrica definida en los
candidatos.}
}
\value{
Un objeto \code{model_stack} ajustado con los pesos óptimos (\code{blended}).
Está listo para el paso final de ajuste a los datos completos (\code{fit_members()}).
}
\description{
Esta función toma los resultados de ajuste de múltiples modelos (candidatos),
los combina en una pila (\code{data_stack}) y ajusta los pesos óptimos
(\code{blend_predictions}) para crear un modelo de ensemble.

Esta función asume que los resultados de ajuste se generaron con
\code{h3sdm_fit_model(..., for_stacking = TRUE)}.
}
\examples{
\dontrun{
# Asumiendo que 'rf_fit' y 'gam_fit' son los resultados de h3sdm_fit_model()
# ensemble_mod <- h3sdm_stack_blend(RF = rf_fit, GAM = gam_fit)

# Paso final de ajuste (Fuera de la función)
# final_ensemble <- fit_members(ensemble_mod)
}
}
\seealso{
Other h3sdm_tools: 
\code{\link{h3sdm_recipe_gam}()},
\code{\link{h3sdm_stack_fit}()},
\code{\link{h3sdm_workflow_gam}()}
}
\concept{h3sdm_tools}
