% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3sdm_stack_fit.R
\name{h3sdm_stack_fit}
\alias{h3sdm_stack_fit}
\title{Creates and fully fits an ensemble model (Stack).}
\usage{
h3sdm_stack_fit(..., non_negative = TRUE, metric = NULL)
}
\arguments{
\item{...}{List objects that are the result of \code{h3sdm_fit_model()}.
Each object must contain the \code{cv_model} element (result of fit_resamples).}

\item{non_negative}{Logical. If \code{TRUE} (default), forces the candidate
model weights to be non-negative.}

\item{metric}{The metric used to optimize the combination of weights.}
}
\value{
A list containing two elements: \code{blended_model} (the stack after blending)
and \code{final_model} (a fully \code{fitted} \code{model_stack} object).
The \code{final_model} is ready for direct prediction with \code{predict()}.
}
\description{
This function combines the process of creating the model stack, optimizing the
weights (\code{blend_predictions}), and fitting the base models to the complete
training set (\code{fit_members()}) into a single step.

\strong{Warning:} It does not follow the canonical tidymodels flow but is convenient.
It requires that the fitting results were generated using \code{h3sdm_fit_model(..., for_stacking = TRUE)}.
}
\seealso{
Other h3sdm_tools: 
\code{\link{h3sdm_recipe_gam}()},
\code{\link{h3sdm_workflow_gam}()}
}
\concept{h3sdm_tools}
