% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3sdm_calculate_it_metrics.R
\name{h3sdm_calculate_it_metrics}
\alias{h3sdm_calculate_it_metrics}
\title{Calculate Information Theory Landscape Metrics for Hexagonal Grid}
\usage{
h3sdm_calculate_it_metrics(landscape_raster, sf_grid)
}
\arguments{
\item{landscape_raster}{A categorical SpatRaster containing land-cover data.}

\item{sf_grid}{An \code{sf} object containing the hexagonal grid with species or land-cover data.}
}
\value{
An \code{sf} object containing the input hex grid with new columns for each calculated metric.
}
\description{
Calculates 5 Information Theory (IT)-based landscape metrics (\code{condent},
\code{ent}, \code{joinent}, \code{mutinf}, \code{relmutinf}) for each hexagon
in a given H3 hexagonal grid.
}
\details{
This function computes landscape metrics using the \code{landscapemetrics::sample_lsm()} workflow.
The results are pivoted to a wide format for easy use.
}
\examples{
\donttest{
# Assuming 'sf_grid' is an sf object with hexagons
# result_sf <- h3sdm_calculate_it_metrics(sf_grid)
# head(result_sf)
}
}
\references{
Hesselbarth et al., 2019. landscapemetrics: an open-source R tool to calculate
landscape metrics. Ecography 42: 1648–1657.

Nowosad & Stepinski, 2019. Information theory as a consistent framework for
landscape patterns. \url{https://doi.org/10.1007/s10980-019-00830-x}
}
