% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3sdm_data.R
\name{h3sdm_data}
\alias{h3sdm_data}
\title{Combina datos para modelos de distribución de especies (SDM) basados en H3}
\usage{
h3sdm_data(pa_sf, predictors_sf)
}
\arguments{
\item{pa_sf}{An sf object from the h3sdm_pa function containing species presence-absence data.}

\item{predictors_sf}{An sf object from the h3sdm_pred function with environmental predictors.}
}
\value{
An sf object containing species, environmental, and spatial data.
}
\description{
Combines species presence-absence data with environmental predictors.
It also calculates the centroid coordinates (x and y) for each hexagon.
}
\examples{
# Assuming you have a species dataset (my_species_pa) and a predictor dataset (my_predictors)
# my_species_pa <- h3sdm_pa(...)
# my_predictors <- h3sdm_predictors(...)
# combined_data <- h3sdm_data(my_species_pa, my_predictors)
}
