% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3sdn_eval_metrics.R
\name{h3sdm_eval_metrics}
\alias{h3sdm_eval_metrics}
\title{Evaluate a single fitted model}
\usage{
h3sdm_eval_metrics(
  fitted_model,
  presence_data = NULL,
  truth_col = "presence",
  pred_col = ".pred_1"
)
}
\arguments{
\item{fitted_model}{The fitted model object.}

\item{presence_data}{A tibble with the true presence locations.}
}
\value{
A tibble with all evaluation metrics.
}
\description{
A helper function to calculate all metrics for a single fitted model.
}
