% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3sdm_fit_model.R
\name{h3sdm_fit_model}
\alias{h3sdm_fit_model}
\title{Fit a tidymodels workflow with spatial data}
\usage{
h3sdm_fit_model(
  sdm_workflow,
  data_split,
  presence_data = NULL,
  truth_col = "presence",
  pred_col = ".pred_1"
)
}
\arguments{
\item{sdm_workflow}{A \code{workflow} object from \code{h3sdm_workflow()}.}

\item{data_split}{An \code{rsplit} object from \code{h3sdm_split()}.}
}
\value{
An \code{rsample_results} object with performance metrics for each fold.
}
\description{
This function trains a tidymodels workflow using a spatial split,
calculating performance metrics for each fold.
}
\examples{
# Assuming a workflow and a data split are defined
# sdm_wf <- h3sdm_workflow(...)
# spatial_split <- h3sdm_split(...)
# fitted_model <- h3sdm_fit(sdm_wf, spatial_split)
}
