% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3sdm_pa.R
\name{h3sdm_pa}
\alias{h3sdm_pa}
\title{Generate presence-absence dataset by H3 hexagons}
\usage{
h3sdm_pa(
  species,
  aoi_sf,
  res = 6,
  n_neg = 500,
  providers = NULL,
  remove_duplicates = FALSE,
  date = NULL,
  expand_factor = 0.1
)
}
\arguments{
\item{species}{Character. Species name (e.g. "Panthera onca")}

\item{aoi_sf}{sf polygon. Area of interest}

\item{res}{Integer. H3 resolution}

\item{n_neg}{Integer. Number of pseudo-absence hexagons to sample}

\item{providers}{Optional. Data providers for get_records()}

\item{remove_duplicates}{Logical. Remove duplicate records}

\item{expand_factor}{Numeric. Buffer factor for grid expansion}
}
\value{
sf object with presence (1/0) column
}
\description{
Generate presence-absence dataset by H3 hexagons
}
